requirements.txt
Django==1.9.5
Fabric==1.11.1
IPy==0.83
Pillow==3.2.0
PyJWT==1.4.0
amqp==1.4.9
anyjson==0.3.3
base32-crockford==0.3.0
beautifulsoup4==4.4.1
billiard==3.3.0.23
bleach==1.4.2
blessings==1.6
boto>=2.40.0
celery==3.1.23
celery-haystack==0.10
celery-with-redis==3.0
certifi==2016.2.28
defusedxml==0.4.1
dj-database-url==0.4.0
django-appconf==1.0.2
django-autoslug==1.9.3
django-cache-url==1.0.0
django-compat==1.0.10
django-compressor==2.0
django-debug-toolbar==0.9.4
django-extensions==1.6.1
django-filter==0.13.0
django-haystack==2.4.1
django-hijack==2.0.3
django-maintenancemode==0.11.2
django-model-utils==2.5
django-nose==1.4.3
django-pylibmc==0.6.1
django-require==1.0.10
django-secure==1.0
django-solo==1.1.2
django-static-sitemaps==4.0.0
django-storages==1.4.1
django-tastypie==0.13.3
ecdsa==0.13
elasticsearch==1.9.0
factory-boy==2.6.0
fake-factory==0.5.7
html2text==3.200.3
html5lib==0.9999999
httplib2==0.9.2
kombu==3.0.35
mock==1.3.0
newrelic==2.64.0.48
nose==1.3.7
nose-exclude==0.4.1
nose-progressive==1.5.1
oauthlib>=1.1
paramiko==1.16.0
pbr==1.9.1
psycopg2==2.6.1
pycrypto==2.6.1
pyelasticsearch==1.4
pylibmc==1.5.0
python-dateutil==2.5.3
python-magic==0.4.11
python-mimeparse==1.5.2
python-social-auth==0.2.18
python-twitter==2.2
python3-openid==3.0.10
pytz==2016.4
rcssmin==1.0.6
redis==2.10.5
requests==2.9.1
requests-oauthlib==0.6.1
rjsmin==1.0.12
simplejson==3.8.2
six==1.10.0
termcolor==1.1.0
urllib3==1.15.1
webcolors==1.5
